cgd1_controller::dialog::display_editor

Struct DisplayEditorWidget

pub struct DisplayEditorWidget {

    pub container: Box,
    pub status_label: Label,
    pub refresh_button: Button,
    brightness_scale: Scale,
    night_brightness_scale: Scale,
    night_mode_switch: Switch,
    screen_duration_spin: SpinButton,
    night_start_entry: TimeEntry,
    night_end_entry: TimeEntry,
    manager: Arc<ClockManager>,
    runtime: Arc<Runtime>,
    connected_address: Arc<Mutex<Option<MacAddress>>>,
    config_store: ConfigStore,

}

Display editor widget for brightness, night brightness, screen timeout, and night mode.

Fields
container: Box

The container box holding all display editor content.

status_label: Label

Status label for operation feedback.

refresh_button: Button

Refresh button.

brightness_scale: Scale

Brightness scale control.

night_brightness_scale: Scale

Night brightness scale control.

night_mode_switch: Switch

Night mode switch.

screen_duration_spin: SpinButton

Screen light duration spin button.

night_start_entry: TimeEntry

Night mode start time entry.

night_end_entry: TimeEntry

Night mode end time entry.

manager: Arc<ClockManager>

Clock manager for device communication.

runtime: Arc<Runtime>

Async runtime.

connected_address: Arc<Mutex<Option<MacAddress>>>

Connected device address.

config_store: ConfigStore

Application config store.

Implementations

impl DisplayEditorWidget

pub fn new( manager: Arc<ClockManager>, runtime: Arc<Runtime>, connected_address: Arc<Mutex<Option<MacAddress>>>, config_store: ConfigStore, ) -> Self

Build the display editor content (without window chrome).

pub fn read_settings(&self)

Read display settings from the connected device and update the UI.