cgd1_controller::window

Struct MainWindow

pub struct MainWindow {

    window: Window,
    runtime: Arc<Runtime>,
    manager: Arc<ClockManager>,
    token_store: Arc<FileTokenStore>,
    connected_address: Arc<Mutex<Option<MacAddress>>>,
    date_display: SevenSegmentDisplay,
    time_display: SevenSegmentDisplay,
    temp_display: SevenSegmentDisplay,
    humidity_display: SevenSegmentDisplay,
    battery_bar: ProgressBar,
    battery_label: Label,
    bluetooth_label: Label,
    device_dropdown: DropDown,
    scan_button: Button,
    connect_button: ToggleButton,
    connect_dot: Box,
    connect_label: Label,
    status_label: Label,
    top_section: Box,
    middle_section: Box,
    bottom_section: Box,

}

The main application window showing the alarm clock display.

Fields
window: Windowruntime: Arc<Runtime>manager: Arc<ClockManager>token_store: Arc<FileTokenStore>connected_address: Arc<Mutex<Option<MacAddress>>>date_display: SevenSegmentDisplaytime_display: SevenSegmentDisplaytemp_display: SevenSegmentDisplayhumidity_display: SevenSegmentDisplaybattery_bar: ProgressBarbattery_label: Labelbluetooth_label: Labeldevice_dropdown: DropDownscan_button: Buttonconnect_button: ToggleButtonconnect_dot: Boxconnect_label: Labelstatus_label: Labeltop_section: Boxmiddle_section: Boxbottom_section: Box
Implementations

impl MainWindow

pub fn new(app: &Application, backend: Backend) -> Self

Create a new main window.

pub fn present(&self)

Present the window.

pub fn window(&self) -> &Window

Get the inner GTK window.

pub fn manager(&self) -> &Arc<ClockManager>

Get the clock manager.

pub fn runtime(&self) -> &Arc<Runtime>

Get the tokio runtime.

pub fn connected_address_arc(&self) -> Arc<Mutex<Option<MacAddress>>>

Get the Arc to the connected address mutex for sharing with closures.

fn setup_layout(&self, _main_box: &Box, top: &Box, middle: &Box, bottom: &Box)

fn setup_signals(&self)

fn setup_resize_handler(&self)

Set up a handler to update font sizes and section proportions when the window is resized.

fn start_clock_tick(&self)

fn start_auto_scan(&self)

Trigger a scan automatically on startup.